+++ components/Alert/react/rollup.config.js
@@ -1,3 +1,13 @@
+import postcss from 'rollup-plugin-postcss'
 import rootRollupConfig from '../../root.rollup.config'
 
-export default rootRollupConfig({ input: './index.ts' })
+export default rootRollupConfig({
+  input: './index.ts',
+  plugins: [
+    postcss({
+      extract: false,
+      module: true,
+      use: ['sass'],
+    }),
+  ],
+})