+++ components/Button/constants.ts
@@ -41,3 +41,26 @@ export const StaticClasses =
 export type ButtonSizes = keyof typeof SizeClassesTable
 
 export type ButtonVariants = keyof typeof VariantClassesTable
+
+export interface ButtonProps {
+  /**
+   * Visual variant to display the button
+   * It will pick colors for font background and border)
+   */
+  variant?: ButtonVariants
+  /**
+   * Size of the button
+   */
+  size?: ButtonSizes
+  /**
+   * Is the button clickable and active?
+   * Note that `variant="disabled"` will also set this
+   */
+  disabled?: boolean
+}
+
+export const ButtonPropsList: Record<keyof ButtonProps, any> = {
+  variant: String,
+  size: String,
+  disabled: Boolean,
+}