+++ icon-registry/src/compileIcon.ts
@@ -1,6 +1,7 @@
-import { type IconProps, type OpenIconProps, iconsMetadata } from './icons'
+import { camelCase } from 'lodash'
+import type { OpenIconProps, IconProps } from './icons'
+import { iconsMetadata, WindiColor } from './icons'
 import { iconSet } from './iconsList'
-import camelCase from 'camelcase'
 
 export const compileIcon = (props: IconProps) => {
   const { name } = props
@@ -24,16 +25,18 @@ export const compileIcon = (props: IconProps) => {
   }
 }
 
-export const getComponentAttributes = ({
-  size,
-  availableSizes,
-  interactiveColorsOnGroup,
-  ...colors
-}: {
-  size: string
-  availableSizes: readonly string[]
-  interactiveColorsOnGroup?: boolean
-} & OpenIconProps) => {
+export const getComponentAttributes = (
+  props: {
+    availableSizes: readonly string[]
+  } & OpenIconProps
+) => {
+  const {
+    size,
+    availableSizes,
+    interactiveColorsOnGroup,
+    name, // not used, just removed from colors
+    ...others
+  } = props
   const sizeWithDefault =
     size ??
     (availableSizes.length >= 1
@@ -42,16 +45,26 @@ export const getComponentAttributes = ({
         : availableSizes[0]
       : '')
 
+  const protectedInteractiveColorsOnGroup =
+    interactiveColorsOnGroup === undefined
+      ? others['interactive-colors-on-group']
+      : interactiveColorsOnGroup
+
+  const colors = others
+
   // TODO: when all icons are converted to using the design system,
   // replace dark by stroke and light by fill,
   // both here and in the windi plugins configs.
   const compiledClasses = Object.keys(colors)
-    .map((color) => {
-      const colorWeight = colors[color]
-      if (!colorWeight) {
+    .map((color: WindiColor) => {
+      if ((color as any) === 'interactive-colors-on-group') {
         return false
       }
-      const lowerCaseColor = color.toLowerCase()
+      const weightedColor = colors[color]
+      if (!weightedColor) {
+        return false
+      }
+      const lowerCaseColor = color.toLowerCase().replace(/-/g, '')
       const colorClass = lowerCaseColor.includes('strokecolor')
         ? 'dark'
         : 'light'
@@ -68,13 +81,13 @@ export const getComponentAttributes = ({
         : ''
 
       const prefix = pseudoClass.length
-        ? interactiveColorsOnGroup !== undefined &&
-          interactiveColorsOnGroup !== false
+        ? protectedInteractiveColorsOnGroup !== undefined &&
+          protectedInteractiveColorsOnGroup !== false
           ? `group-${pseudoClass}:`
           : `icon-${pseudoClass}:`
         : ''
 
-      const finalClass = `${prefix}icon-${colorClass}${secondaryClass}-${colorWeight}`
+      const finalClass = `${prefix}icon-${colorClass}${secondaryClass}-${weightedColor}`
 
       return finalClass
     })