+++ test/vue-app/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @cypress-design/test-vue-app
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [[`2fdaf6b`](https://github.com/cypress-io/cypress-design/commit/2fdaf6be6f81a2a851761258347ed213577c5b26)]:
+  - @cypress-design/vue-button@0.2.0
+  - @cypress-design/vue-checkbox@0.3.0
+
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [[`e3eb9c7`](https://github.com/cypress-io/cypress-design/commit/e3eb9c7fee2d7a6e0a773e85ed4b73be04d83587)]:
+  - @cypress-design/vue-alert@0.0.2
+  - @cypress-design/vue-checkbox@0.2.1
+  - @cypress-design/vue-icon@0.4.2
+
 ## 0.0.6
 
 ### Patch Changes